Modifying SPI Driver & Solving Buffer Transmission Error

+++ MCAL/Inc/SPI_Interface.h
@@ -9,7 +9,7 @@
 
 /******************* MAIN INCLUDES *********************/
 #include <stdint.h>
-#include "../../LIBRARY//ErrTypes.h"
+#include "../../LIBRARY/ErrTypes.h"
 /*******************************************************/
 
 
@@ -168,7 +168,7 @@ Error_State_t SPI_Receive(const SPI_CONFIGS_t * SPI_Config, uint16_t * Received_
  * @param			:	CallBack Function
  * @retval			:	Error State
  */
-Error_State_t SPI_Transmit_IT(const SPI_CONFIGS_t * SPI_Config, uint16_t * Data , uint8_t Buffer_Size, void (* SPI_TXC_CallBackFunc)(void));
+Error_State_t SPI_Transmit_IT(const SPI_CONFIGS_t * SPI_Config, uint8_t * Data , uint8_t Buffer_Size, void (* SPI_TXC_CallBackFunc)(void));
 
 /*
  * @function 		:	SPI_Receive_IT
@@ -179,7 +179,7 @@ Error_State_t SPI_Transmit_IT(const SPI_CONFIGS_t * SPI_Config, uint16_t * Data 
  * @param			:	CallBack Function
  * @retval			:	Error State
  */
-Error_State_t SPI_Receive_IT(const SPI_CONFIGS_t * SPI_Config, uint16_t * Received_Data ,uint8_t Buffer_Size , void (* SPI_RXC_CallBackFunc)(void));
+Error_State_t SPI_Receive_IT(const SPI_CONFIGS_t * SPI_Config, uint8_t * Received_Data ,uint8_t Buffer_Size , void (* SPI_RXC_CallBackFunc)(void));
 
 /*
  * @function 		:	SPI_Enable_DMA_RX
